refactor: extract ResourceManagerView.kt
